refactor(components): Move app-authored components out of ui into custom
Changes:
- Add custom/animation (animated-group, reveal-elements, scroll-reveal, stagger-children)
- Add custom/form (multi-select, password-input, stepper-number-input)
- Move banner, bg-animate-button and remove-icon-button into custom/common
- Repoint every import and I-prefix the five Props interfaces that lacked it

components/ui is now stock shadcn only, apart from button and badge which the
project deliberately owns, so a shadcn upgrade can no longer collide with app
code.

Closes #50
