+++ frontend/src/app/(root)/components/sections/final-cta-section.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { Button } from "@/components/ui/button";
-import { ScrollReveal } from "@/components/ui/scroll-reveal";
+import { ScrollReveal } from "@/components/custom/animation/scroll-reveal";
 import SparkleField from "@/app/(root)/components/doodles/sparkle-field";
 import CartLogo from "@/components/icons/cart-logo";
