+++ frontend/src/app/(root)/components/sections/pricing-section.tsx
@@ -1,7 +1,7 @@
 import ComingSoonBadge from "@/components/custom/common/coming-soon-badge";
 import SectionHeading from "@/app/(root)/components/sections/section-heading";
 import PricingReceipt from "@/app/(root)/components/sections/pricing-receipt";
-import { ScrollReveal } from "@/components/ui/scroll-reveal";
+import { ScrollReveal } from "@/components/custom/animation/scroll-reveal";
 import { freePlanRows, proPlanRows } from "@/app/(root)/data/landing";
 
 export default function PricingSection() {
