+++ frontend/src/app/products/components/forms/watchlist-item-modal.tsx
@@ -6,13 +6,13 @@ import { Eye, Save, TriangleAlert } from "lucide-react";
 import { ModalShell } from "@/components/custom/modal/modal-shell";
 import { Form } from "@/components/ui/form";
 import { Skeleton } from "@/components/ui/skeleton";
-import { RemoveIconButton } from "@/components/ui/remove-icon-button";
+import { RemoveIconButton } from "@/components/custom/common/remove-icon-button";
 import { WatchType } from "@/lib/api";
 import cijeneService from "@/lib/cijene-api";
 import { closeModalUrl } from "@/lib/modal/modal-navigation";
 import type { WatchTypeParam } from "@/lib/modal/modal-registry";
 import ProductInfoDisplay from "@/app/products/components/product-info-display";
-import { Banner } from "@/components/ui/banner";
+import { Banner } from "@/components/custom/common/banner";
 import {
   getAveragePrice,
   getMinPrice,
