+++ frontend/src/components/custom/animation/reveal-elements.ts
