+++ frontend/src/components/custom/animation/stagger-children.tsx
@@ -2,54 +2,37 @@
 
 import { Children, ReactNode } from "react";
 import { motion } from "motion/react";
-import type { Variants } from "motion/react";
 import { useReducedMotionSafe } from "@/hooks/use-reduced-motion-safe";
 import {
   itemTagFor,
   type RevealContainerTag,
   type RevealItemTag,
-} from "@/components/ui/reveal-elements";
+} from "@/components/custom/animation/reveal-elements";
 
-type RevealPreset = "rise" | "pop" | "swing";
-
-interface IScrollRevealProps {
+interface IStaggerChildrenProps {
   children: ReactNode;
   className?: string;
-  preset?: RevealPreset;
+  // Vertical travel of each item as it fades in.
+  distance?: number;
+  // Delay between consecutive items.
   stagger?: number;
-  amount?: number;
+  duration?: number;
   as?: RevealContainerTag;
   itemAs?: RevealItemTag;
 }
 
-const spring = { type: "spring", stiffness: 300, damping: 18 } as const;
-
-const presets: Record<RevealPreset, Variants> = {
-  rise: {
-    hidden: { opacity: 0, y: 32 },
-    visible: { opacity: 1, y: 0, transition: spring },
-  },
-  pop: {
-    hidden: { opacity: 0, scale: 0.85 },
-    visible: { opacity: 1, scale: 1, transition: spring },
-  },
-  swing: {
-    hidden: { opacity: 0, y: 24, rotate: -3 },
-    visible: { opacity: 1, y: 0, rotate: 0, transition: spring },
-  },
-};
-
-// Scroll-triggered staggered reveal: copy stays server-rendered because it
-// crosses this client boundary as children. No-op under reduced motion.
-export function ScrollReveal({
+// Staggered fade + rise reveal for modal/section content; no-op under reduced
+// motion (hydration-safe: the plain-element branch only activates after mount).
+// Accepts a single child or an array (falsy children are dropped).
+export function StaggerChildren({
   children,
   className,
-  preset = "rise",
-  stagger = 0.1,
-  amount = 0.3,
+  distance = 8,
+  stagger = 0.05,
+  duration = 0.25,
   as = "div",
   itemAs,
-}: IScrollRevealProps) {
+}: IStaggerChildrenProps) {
   const reduceMotion = useReducedMotionSafe();
   const items = Children.toArray(children);
   const itemTag = itemTagFor(as, itemAs);
@@ -74,12 +57,21 @@ export function ScrollReveal({
     <Container
       className={className}
       initial="hidden"
-      whileInView="visible"
-      viewport={{ once: true, amount }}
+      animate="visible"
       variants={{ visible: { transition: { staggerChildren: stagger } } }}
     >
       {items.map((child, index) => (
-        <Item key={index} variants={presets[preset]}>
+        <Item
+          key={index}
+          variants={{
+            hidden: { opacity: 0, y: distance },
+            visible: {
+              opacity: 1,
+              y: 0,
+              transition: { duration, ease: "easeOut" },
+            },
+          }}
+        >
           {child}
         </Item>
       ))}
