+++ frontend/src/components/custom/auth/auth-modal.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { CircleAlert } from "lucide-react";
 
 import { ModalShell } from "@/components/custom/modal/modal-shell";
-import { Banner } from "@/components/ui/banner";
+import { Banner } from "@/components/custom/common/banner";
 import ForgotPasswordForm from "@/components/custom/auth/components/forms/forgot-password-form";
 import AuthModeSwitch from "@/components/custom/auth/components/auth-mode-switch";
 import AuthCredentialsPanel from "@/components/custom/auth/components/forms/auth-credentials-panel";
