+++ frontend/src/components/custom/common/banner.tsx
