+++ frontend/src/components/custom/settings/tabs/preferences-tab.tsx
@@ -2,7 +2,7 @@
 
 import { Store, MapPin } from "lucide-react";
 
-import { StaggerChildren } from "@/components/ui/stagger-children";
+import { StaggerChildren } from "@/components/custom/animation/stagger-children";
 import SettingsSection from "@/components/custom/settings/ui/settings-section";
 import PinnedStoresGrid from "@/components/custom/settings/components/pinned-stores-grid";
 import PinnedPlacesSelect from "@/components/custom/settings/components/pinned-places-select";
