+++ frontend/src/components/custom/settings/tabs/security-tab.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { StaggerChildren } from "@/components/ui/stagger-children";
+import { StaggerChildren } from "@/components/custom/animation/stagger-children";
 import CredentialsForm from "@/components/custom/settings/security/components/credentials-form";
 import LinkedAccounts from "@/components/custom/settings/security/components/linked-accounts";
 import AccountActions from "@/components/custom/settings/security/components/account-actions";
