+++ frontend/src/components/custom/store-chain/store-chain-multi-select.tsx
@@ -5,7 +5,7 @@ import {
   MultiSelectItem,
   MultiSelectTrigger,
   MultiSelectValue,
-} from "@/components/ui/multi-select";
+} from "@/components/custom/form/multi-select";
 import { getChainLabel } from "@/utils/labels";
 import { compareHr } from "@/utils/strings";
